+++ src/Atata.GlobalFixture/Class.cs
@@ -1,0 +1,12 @@
+﻿namespace $rootnamespace$;
+
+public sealed class $safeitemrootname$ : AtataGlobalFixture
+{
+	protected override void ConfigureAtataContextBaseConfiguration(AtataContextBuilder builder)
+    {
+    }
+
+    protected override void ConfigureGlobalAtataContext(AtataContextBuilder builder)
+    {
+    }
+}