+++ src/Atata.GlobalFixture/Properties/AssemblyInfo.cs
@@ -1,6 +1,6 @@
 ﻿using System.Reflection;
 using System.Runtime.InteropServices;
 
-[assembly: AssemblyTitle("Atata.NUnitBaseTestFixture")]
+[assembly: AssemblyTitle("Atata.GlobalFixture")]
 [assembly: AssemblyDescription("")]
 [assembly: Guid("23e1eb9d-6ec3-4acd-afa1-e3eb41ee10ca")]